% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sg_gps.R
\name{read_sg_gps}
\alias{read_sg_gps}
\title{Read in Sensorgnome GPS Data}
\usage{
read_sg_gps(x, tz = "UTC")
}
\arguments{
\item{x}{the file name and path of the .txt file you want to read.}

\item{keep_only_ports}{logical. Specifies if you want to filter out all data except pulse data. Defaults to TRUE. It is highly recommended that you do not change this parameter unless you have experience working with this data.}
}
\description{
This function reads the complied .txt file created by the \code{\link[=extract_sg]{extract_sg()}} function and outputs a data fram with just the Sensorgnome GPS data, which is useful for diagnosing issues.
}
